backend/requirements.txt
pydantic>=2.0
numpy
fastapi
uvicorn
# supabase
python-dotenv
sqlmodel
psycopg2-binary
alembic